Extract methods to a new class ModuleLoader
