+++ manimlib/config.py
@@ -12,6 +12,7 @@ import sys
 import yaml
 
 from manimlib.logger import log
+from manimlib.module_loader import ModuleLoader
 from manimlib.utils.dict_ops import merge_dicts_recursively
 from manimlib.utils.init_config import init_customization
 
@@ -205,81 +206,6 @@ def get_manim_dir():
     return os.path.abspath(os.path.join(manimlib_dir, ".."))
 
 
-def exec_module_and_track_imports(spec, module: Module):
-    """
-    Executes the given module (imports it) and returns all the modules that
-    are imported during its execution.
-
-    This is achieved by replacing the __import__ function with a custom one that
-    tracks the imported modules. At the end, the original __import__ function is
-    restored.
-    """
-    imported_modules = set()
-    original_import = builtins.__import__
-
-    def tracked_import(name, globals=None, locals=None, fromlist=(), level=0):
-        result = original_import(name, globals, locals, fromlist, level)
-        imported_modules.add(name)
-        return result
-
-    builtins.__import__ = tracked_import
-
-    try:
-        spec.loader.exec_module(module)
-    finally:
-        builtins.__import__ = original_import
-
-    return imported_modules
-
-
-def reload_modules(modules, reloaded_modules_tracker: set):
-    """
-    Out of the given modules, reloads the ones that were not already imported.
-
-    We also restrict ourselves to reloading only the modules that are not
-    built-in Python modules to avoid potential issues since they were mostly
-    not designed to be reloaded.
-    """
-    for mod in modules:
-        if mod in reloaded_modules_tracker:
-            continue
-
-        if mod not in sys.modules:
-            continue
-
-        if mod in sys.builtin_module_names or mod in ["pkg_resources", "setuptools"]:
-            continue
-
-        log.debug('Reloading module "%s"', mod)
-        importlib.reload(sys.modules[mod])
-
-        reloaded_modules_tracker.add(mod)
-
-
-def get_module(file_name: str | None, is_during_reload=False) -> Module | None:
-    """
-    Imports a module from a file and returns it.
-
-    During reload (when the user calls `reload()` in the IPython shell), we
-    also track the imported modules and reload them as well (they would be
-    cached otherwise). See the reload_manager where the reload parameter is set.
-    """
-    if file_name is None:
-        return None
-    module_name = file_name.replace(os.sep, ".").replace(".py", "")
-    spec = importlib.util.spec_from_file_location(module_name, file_name)
-    module = importlib.util.module_from_spec(spec)
-
-    if is_during_reload:
-        imported_modules = exec_module_and_track_imports(spec, module)
-        reloaded_modules_tracker = set()
-        reload_modules(imported_modules, reloaded_modules_tracker)
-    else:
-        spec.loader.exec_module(module)
-
-    return module
-
-
 def get_indent(line: str):
     return len(line) - len(line.lstrip())
 
@@ -347,7 +273,7 @@ def get_module_with_inserted_embed_line(
     with open(new_file, 'w') as fp:
         fp.writelines(new_lines)
 
-    module = get_module(new_file, is_during_reload)
+    module = ModuleLoader.get_module(new_file, is_during_reload)
     # This is to pretend the module imported from the edited lines
     # of code actually comes from the original file.
     module.__file__ = file_name
@@ -359,7 +285,7 @@ def get_module_with_inserted_embed_line(
 
 def get_scene_module(args: Namespace) -> Module:
     if args.embed is None:
-        return get_module(args.file)
+        return ModuleLoader.get_module(args.file)
     else:
         return get_module_with_inserted_embed_line(
             args.file, args.scene_names[0], args.embed, args.is_reload
